requirements.txt
coverage
coveralls
wheel
premailer
django-annoying
six
setuptools
nose
